[improve] move to api model
